+++ src/app/mate/[id]/components/MateInfo/components/MateActions/MateActions.tsx
@@ -4,14 +4,12 @@ import * as S from "./MateActions.styles";
 
 import { useMemo } from "react";
 
-import { useCancelParticipant, useRequestParticipant } from "@/app/mate/[id]/hooks";
-import { useUI } from "@/components/uiContext/UiContext";
 import useAuthStore from "@/lib/stores/useAuthStore";
 import useMeStore from "@/lib/stores/useMeStore";
 import { Participant } from "@/types/OriginDataType";
 import { MateDetailResponse } from "@/types/Response";
 
-import { MateButton } from "./components";
+import { useMateButtonList } from "./hooks";
 
 interface MateActionsProps {
   mateData: MateDetailResponse;
@@ -20,15 +18,8 @@ interface MateActionsProps {
 }
 
 const MateActions = ({ mateData, acceptedUserList, requestedUserList }: MateActionsProps) => {
-  const { setModalView, openModal, closeModal } = useUI();
   const { userId } = useMeStore();
   const { isLogIn } = useAuthStore();
-
-  const cancelParticipantMutation = useCancelParticipant({
-    successMessage: "정상적으로 메이트에서 탈퇴되었습니다.",
-  });
-
-  const postParticipantRequestMutation = useRequestParticipant();
 
   const participantId = useMemo(() => {
     if (!userId) {
@@ -45,6 +36,11 @@ const MateActions = ({ mateData, acceptedUserList, requestedUserList }: MateActi
 
     return false;
   }, [mateData, userId]);
+
+  const MateButtons = useMateButtonList({
+    participantId: participantId,
+    mateId: mateData.id,
+  });
 
   const userStatus = useMemo(() => {
     if (!isLogIn) {
@@ -74,106 +70,24 @@ const MateActions = ({ mateData, acceptedUserList, requestedUserList }: MateActi
     return "NO_ACTION";
   }, [requestedUserList, acceptedUserList, mateData, userId]);
 
-  const handleClickRequest = () => {
-    setModalView("MATE_REQUEST_VIEW");
-    openModal({
-      onAccept: (message: string) => {
-        postParticipantRequestMutation.mutate({ message, mateId: mateData.id });
-        closeModal();
-      },
-    });
-  };
-
-  const handleClickCancel = () => {
-    if (!participantId) {
-      return;
-    }
-
-    setModalView("CONFIRM_VIEW");
-    openModal({
-      message: "정말로 메이트에서 탈퇴하시겠어요?",
-      warningMessage: "나가시면 되돌리실 수 없습니다.",
-      onClickAccept: () => {
-        cancelParticipantMutation.mutate({ mateId: mateData.id, participantId: participantId });
-      },
-    });
-  };
-
-  const handleClickAccessLogin = () => {
-    setModalView("ACCESS_LOGIN_VIEW");
-    openModal();
-  };
-
-  const ButtonList = {
-    Login: (
-      <MateButton
-        text="로그인 후 이용하실 수 있습니다."
-        onClick={handleClickAccessLogin}
-      />
-    ),
-
-    Request: (
-      <MateButton
-        text="메이트 신청하기"
-        onClick={handleClickRequest}
-      />
-    ),
-
-    Pending: (
-      <MateButton
-        text="요청중"
-        disabled
-      />
-    ),
-
-    Cancel: (
-      <MateButton
-        text="취소하기"
-        onClick={handleClickCancel}
-        isSecondButton
-      />
-    ),
-
-    Accepted: (
-      <MateButton
-        text="현재 참여중 입니다."
-        disabled
-      />
-    ),
-
-    Completed: (
-      <MateButton
-        text="종료된 메이트입니다."
-        disabled
-      />
-    ),
-
-    Author: (
-      <MateButton
-        text="메이트 모집중..."
-        disabled
-      />
-    ),
-  };
-
   return (
     <S.MateActionsLayout>
-      {userStatus === "NOT_LOGIN" && ButtonList.Login}
-      {userStatus === "NO_ACTION" && ButtonList.Request}
+      {userStatus === "NOT_LOGIN" && MateButtons.Login}
+      {userStatus === "NO_ACTION" && MateButtons.Request}
       {userStatus === "REQUESTED" && (
         <>
-          {ButtonList.Pending}
-          {ButtonList.Cancel}
+          {MateButtons.Pending}
+          {MateButtons.Cancel}
         </>
       )}
       {userStatus === "ACCEPTED" && (
         <>
-          {ButtonList.Accepted}
-          {ButtonList.Cancel}
+          {MateButtons.Accepted}
+          {MateButtons.Cancel}
         </>
       )}
-      {userStatus === "CLOSE" && <>{ButtonList.Completed}</>}
-      {userStatus === "AUTHOR" && ButtonList.Author}
+      {userStatus === "CLOSE" && <>{MateButtons.Completed}</>}
+      {userStatus === "AUTHOR" && MateButtons.Author}
     </S.MateActionsLayout>
   );
 };
