+++ src/app/mate/[id]/components/MateInfo/components/MateActions/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useMateButtonList } from "./useMateButtonList/useMateButtonList";
